Move logic for outside clicks to AddList component
