+++ src/components/ListEditor.js
@@ -13,30 +13,6 @@ const ListEditor = (props) => {
     }
   };
 
-  // Custom hook for detecting click outside of the element
-  const useOnClickOutside = (ref, handler) => {
-    useEffect(() => {
-      const listener = (event) => {
-        // Do nothing if clicking ref's element or descendent elements
-        if (!ref.current || ref.current.contains(event.target)) {
-          return;
-        }
-
-        handler(event);
-      };
-
-      document.addEventListener('mousedown', listener);
-      document.addEventListener('touchstart', listener);
-      return () => {
-        document.removeEventListener('mousedown', listener);
-        document.removeEventListener('touchstart', listener);
-      };
-    }, [ref, handler]);
-  };
-
-  // Calling hook
-  useOnClickOutside(ref, () => props.onClickOutside(false));
-
   return (
     <div className={styles.ListTitleEdit} ref={ref}>
       <TextareaAutosize
